requirements.txt
appdirs==1.4.3
attrs==18.1.0
beautifulsoup4==4.6.3
biopython==1.71
biothings-client==0.1.1
bottle==0.12.13
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
CProfileV==1.0.7
cycler==0.10.0
decorator==4.3.0
future==0.16.0
h2o==3.18.0.8
html5lib==1.0.1
idna==2.6
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
MechanicalSoup==0.10.0
more-itertools==4.1.0
mygene==3.0.0
nbformat==4.4.0
nose==1.3.7
numpy==1.14.2
pandas==0.23.4
plotly==2.5.1
pluggy==0.6.0
py==1.5.3
PyEntrezId==1.5.8.2
pyfasta==0.5.2
Pygments==2.1.3
pyparsing==2.2.0
pytest==3.5.1
python-dateutil==2.7.2
python-termstyle==0.1.10
pytools==2018.4
pytz==2018.4
rednose==1.1.1
requests==2.10.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tabulate==0.8.2
traitlets==4.3.2
urllib3==1.22
webencodings==0.5.1
xmldict==0.4.1
xmltodict==0.10.1
